Fixed a few things
