+++ Analysis/rnaseq.normalize_splice_junction_values.pbs
@@ -5,4 +5,24 @@
 #PBS -d .
 #PBS -l vmem=10g,mem=10g
 
+if [ -z $splice_file];
+then
+	splice_file=$1
+fi
+
+if [ -z $transcript_model];
+then
+	transcript_model=$2
+fi
+
+if [ -z $action];
+then
+	action=$3
+fi
+
+if [ -z $outputFileName];
+then
+	outputFileName=$4
+fi
+
 echo ~/tools/MendelianRNA-seq/Analysis/NormalizeSpliceJunctionValues.py -splice_file $splice_file -transcript_model $transcript_model $action > $outputFileName
